create symlinks to ../ld/ in the linkers so can build all of them
Before that we would get linking errors when building in sequence
the linkers because one ../ld/elf.o was built for 6l and then
reused when building 8l but the l.h was different so the object too
should be difference.
Simpler to use symlinks.
Unfortunately symlinks are not supported in plan9 so long term we need
to find a different solution
maybe like for lexbody and macbody and just include