Move pgen.c and pswt.c to cc2/
Those were not part of the cc lib in cc/ so clearer
to separate them. Those files actually depend on gc.h
defined by each compiler and must be recompiled
for each compiler so you should not use ../cc/pswt.$O
because it will not work because you must recompile
it for each compiler.

This was actually causing some CI regression and segfault in 6l
when we were changing the order of dirs to compile because
the objects were not compiled again.