import 5a arm assembler (renamed 5a_) from https://bitbucket.org/plan9-from-bell-labs/9-cc.git
I renamed it 5a_ to avoid conflict with the go-extended 5a.

The rational for importing again 5a/5c/5l
is that this original version might be simpler than
the go-extended one; For example it still uses setR12
and SB so the distance to my books in principia is smaller.
Moreover with 7a/7c/7l I recently saw it was doable
to import the original assembler and make them (half)
work on Linux by adjusting a few ELF headers.