long -> int32 for linkers/*/x.out.h
This solves the segfault when linking principia/compilers/hoc
which happened probably when 8l was linking libc.a(sin.c)
which contained float and because of the use of sprint()
in 8l/obj.c sprint() probably overflowed and corrupted the memory.

Note that we should get rid of sprint too!