mk: use double for node time using new libc mtime_ field
We need subsecond resolution for comparing the time of file because
on my thelioastra arm64 with NPROC=64, many object files can be
built in the same second and this can then confuse mk and
trigger too many recompilation a second time. also when I was using
'<' instead of '<=' for time comparison, this would also lead to
missing running certain commands such as 'iar' to add new
objects to the libc.a in one dir because the objects in
the dir were made in the same second than the libc.a modified in
the previous dir.

=> subsecond resolution for mk.